requirements.txt
python-twitch-client
praw
spotipy
wikipedia
discord
serial
geocoder
forecastiopy